Preprocessor now outputs PPCharacters. User can get the output character and its offset to the CodeBuffer from this wrapper class.
--

Wrote a simple lookahead buffer for the Preprocessor
and implemented the `peek()` and `reachedEndOfInput()`
function upon it.

--

Previously `reachedEndOfInput` cannot sense the
scan has finished if the input contains a macro
that expand to nothing at the end of input.

For example:

```c
#define ID(x) x
a empty id: ID()
```

It happened because the function determine if
the scan is finished by examinating the internal
scanner's current offset. Clearly before expanding
the `ID()` macro it hasn't reached the end of the
input string, but since `ID()` expands to nothing,
there is no more content to be output, and hence
the scanning is done, yet the function cannot
know it.

This problem is now fixed by implementing the
`peek()` function and let the function examinates
if the `peek()` function returns EOF.

--

Simplified the PPImpl's constructor.