[wip] fixed program crashing.
Reason:

When I call `enterSecton` to enter a zero-length section, the function
didn't check the length and blindly pushed the id and offset to the stacks,
and the function  `get()` assumes the `_offset` is valid when it's called,
and will not check if the `_offset` has reached the end. However it isn't
true if the section's length is zero, so when I `get()` a character it
will actually try reading characters that are next to the end of the
section, causing a out of index access and the program crashes.

Solution:

Add a guarding `if` to the `enterSection`, the the pushing section has
zero length, don't do anything, just return.