[wip] Solved two problems.
# Problem 1

The evaluatation order wasn't correct, for example:

```c
#define EMPTY()
#define DELAY(macro) macro EMPTY()
#define ID(x) x

DELAY(ID)(3)
```
First the preprocessor fully expanded DELAY(ID) to a string "ID ", causing
`DELAY(ID)(3)` becomes `ID (3)`, then it discovered an function-like macro
and expanded it to `3`. This is not correct.

The correct order is, after replacing the parameters with arguments, it
should just replace the macro expression with the macro body. `DELAY(ID)(3)`
will become `ID EMPTY (3)`, then continue scanning, resulting the final
string `ID (3)`.

# Problem 2

If there is a macro with empty body between two identifiers, and it's
surrounded by spaces, a extra space is outputed.

For example:

```c
#define EMPTY()
#define A() a
#definb B() b

A() EMPTY() B()
````

will produces `a  b`, which is not correct. the correct result should be
`a b`, with only one space between two identifiers.